+++ app/views/ui/panel.js
